+++ ext/torch/device.cpp
@@ -1,11 +1,15 @@
+#include <string>
+
 #include <torch/torch.h>
 
 #include <rice/rice.hpp>
+#include <rice/stl.hpp>
 
 #include "utils.h"
 
 void init_device(Rice::Module& m) {
-  Rice::define_class_under<torch::Device>(m, "Device")
+  auto rb_cDevice = Rice::define_class_under<torch::Device>(m, "Device");
+  rb_cDevice
     .define_constructor(Rice::Constructor<torch::Device, const std::string&>())
     .define_method(
       "_index",
@@ -25,8 +29,10 @@ void init_device(Rice::Module& m) {
         return s.str();
       })
     .define_method(
-      "_str",
+      "to_s",
       [](torch::Device& self) {
         return self.str();
       });
+
+  THPDeviceClass = rb_cDevice.value();
 }