+++ ext/torch/ruby_arg_parser.cpp
@@ -1,7 +1,13 @@
 // adapted from PyTorch - python_arg_parser.cpp
 
+#include <string>
+#include <unordered_map>
+#include <unordered_set>
+#include <vector>
+
 #include "ruby_arg_parser.h"
 
+VALUE THPDeviceClass = Qnil;
 VALUE THPGeneratorClass = Qnil;
 VALUE THPVariableClass = Qnil;
 
@@ -99,7 +105,7 @@ FunctionParameter::FunctionParameter(const std::string& fmt, bool keyword_only)
   ruby_name = THPUtils_internSymbol(name);
   auto np_compat_it = numpy_compatibility_arg_names.find(name);
   if (np_compat_it != numpy_compatibility_arg_names.end()) {
-    for (const auto& str: np_compat_it->second) {
+    for (const auto& str : np_compat_it->second) {
       numpy_python_names.push_back(THPUtils_internSymbol(str));
     }
   }
@@ -190,8 +196,7 @@ static bool is_int_or_symint_list(VALUE obj, int broadcast_size) {
 }
 
 // argnum is needed for raising the TypeError, it's used in the error message.
-auto FunctionParameter::check(VALUE obj, int argnum) -> bool
-{
+auto FunctionParameter::check(VALUE obj, int argnum) -> bool {
   switch (type_) {
     case ParameterType::TENSOR: {
       if (THPVariable_Check(obj)) {
@@ -253,7 +258,7 @@ auto FunctionParameter::check(VALUE obj, int argnum) -> bool
     case ParameterType::LAYOUT: return SYMBOL_P(obj);
     case ParameterType::MEMORY_FORMAT: return false; // return THPMemoryFormat_Check(obj);
     case ParameterType::QSCHEME: return false; // return THPQScheme_Check(obj);
-    case ParameterType::DEVICE: return RB_TYPE_P(obj, T_STRING); // TODO check device
+    case ParameterType::DEVICE: return RB_TYPE_P(obj, T_STRING) || THPDevice_Check(obj);
     case ParameterType::STRING: return RB_TYPE_P(obj, T_STRING);
     case ParameterType::SYM_INT: return is_int_or_symint(obj);
     case ParameterType::SYM_INT_LIST: return is_int_or_symint_list(obj, size);