+++ ext/torch/ruby_arg_parser.h
@@ -3,6 +3,9 @@
 #pragma once
 
 #include <sstream>
+#include <unordered_map>
+#include <string>
+#include <vector>
 
 #include <torch/torch.h>
 #include <rice/rice.hpp>
@@ -162,7 +165,7 @@ inline std::array<at::Tensor, N> RubyArgs::tensorlist_n(int i) {
   Check_Type(arg, T_ARRAY);
   auto size = RARRAY_LEN(arg);
   if (size != N) {
-    rb_raise(rb_eArgError, "expected array of %d elements but got %d", N, (int)size);
+    rb_raise(rb_eArgError, "expected array of %d elements but got %d", N, static_cast<int>(size));
   }
   for (int idx = 0; idx < size; idx++) {
     VALUE obj = rb_ary_entry(arg, idx);
@@ -354,8 +357,11 @@ inline at::Device RubyArgs::device(int i) {
   if (NIL_P(args[i])) {
     return at::Device("cpu");
   }
-  const std::string &device_str = THPUtils_unpackString(args[i]);
-  return at::Device(device_str);
+  if (RB_TYPE_P(args[i], T_STRING)) {
+    const std::string &device_str = THPUtils_unpackString(args[i]);
+    return at::Device(device_str);
+  }
+  return Rice::detail::From_Ruby<at::Device>().convert(args[i]);
 }
 
 inline at::Device RubyArgs::deviceWithDefault(int i, const at::Device& default_device) {
@@ -463,7 +469,7 @@ struct RubyArgParser {
     template<int N>
     inline RubyArgs parse(VALUE self, int argc, VALUE* argv, ParsedArgs<N> &dst) {
       if (N < max_args) {
-        rb_raise(rb_eArgError, "RubyArgParser: dst ParsedArgs buffer does not have enough capacity, expected %d (got %d)", (int)max_args, N);
+        rb_raise(rb_eArgError, "RubyArgParser: dst ParsedArgs buffer does not have enough capacity, expected %d (got %d)", static_cast<int>(max_args), N);
       }
       return raw_parse(self, argc, argv, dst.args);
     }