+++ ext/torch/templates.h
@@ -1,12 +1,12 @@
 #pragma once
 
+#include <string>
+
 #ifdef isfinite
 #undef isfinite
 #endif
 
 #include <rice/rice.hpp>
-
-using namespace Rice;
 
 using torch::Device;
 using torch::Scalar;
@@ -31,7 +31,7 @@ using torch::nn::init::NonlinearityType;
 
 #define END_HANDLE_TH_ERRORS                                         \
   } catch (const torch::Error& ex) {                                 \
-    rb_raise(rb_eRuntimeError, "%s", ex.what_without_backtrace());   \
+    rb_raise(rb_eTorchError, "%s", ex.what_without_backtrace());   \
   } catch (const Rice::Exception& ex) {                              \
     rb_raise(ex.class_of(), "%s", ex.what());                        \
   } catch (const std::exception& ex) {                               \
@@ -41,57 +41,61 @@ using torch::nn::init::NonlinearityType;
 #define RETURN_NIL                                                   \
   return Qnil;
 
-namespace Rice::detail
-{
+namespace Rice::detail {
   template<typename T>
-  struct Type<c10::complex<T>>
-  {
-    static bool verify()
-    {
-      return true;
-    }
+  struct Type<c10::complex<T>> {
+    static bool verify() { return true; }
   };
 
   template<typename T>
-  class To_Ruby<c10::complex<T>>
-  {
+  class To_Ruby<c10::complex<T>> {
   public:
-    VALUE convert(c10::complex<T> const& x)
-    {
+    To_Ruby() = default;
+
+    explicit To_Ruby(Arg* arg) : arg_(arg) { }
+
+    VALUE convert(c10::complex<T> const& x) {
       return rb_dbl_complex_new(x.real(), x.imag());
     }
+
+  private:
+    Arg* arg_ = nullptr;
   };
 
   template<typename T>
-  class From_Ruby<c10::complex<T>>
-  {
+  class From_Ruby<c10::complex<T>> {
   public:
-    c10::complex<T> convert(VALUE x)
-    {
+    From_Ruby() = default;
+
+    explicit From_Ruby(Arg* arg) : arg_(arg) { }
+
+    double is_convertible(VALUE value) { return Convertible::Exact; }
+
+    c10::complex<T> convert(VALUE x) {
       VALUE real = rb_funcall(x, rb_intern("real"), 0);
       VALUE imag = rb_funcall(x, rb_intern("imag"), 0);
       return c10::complex<T>(From_Ruby<T>().convert(real), From_Ruby<T>().convert(imag));
     }
-  };
-}
 
-namespace Rice::detail
-{
-  template<>
-  struct Type<FanModeType>
-  {
-    static bool verify()
-    {
-      return true;
-    }
+  private:
+    Arg* arg_ = nullptr;
   };
 
   template<>
-  class From_Ruby<FanModeType>
-  {
+  struct Type<FanModeType> {
+    static bool verify() { return true; }
+  };
+
+  template<>
+  class From_Ruby<FanModeType> {
   public:
-    FanModeType convert(VALUE x)
-    {
+    From_Ruby() = default;
+
+    explicit From_Ruby(Arg* arg) : arg_(arg) { }
+
+    double is_convertible(VALUE value) { return Convertible::Exact; }
+
+    FanModeType convert(VALUE x) {
       auto s = String(x).str();
       if (s == "fan_in") {
         return torch::kFanIn;
@@ -101,23 +105,26 @@ namespace Rice::detail
         throw std::runtime_error("Unsupported nonlinearity type: " + s);
       }
     }
+
+  private:
+    Arg* arg_ = nullptr;
   };
 
   template<>
-  struct Type<NonlinearityType>
-  {
-    static bool verify()
-    {
-      return true;
-    }
+  struct Type<NonlinearityType> {
+    static bool verify() { return true; }
   };
 
   template<>
-  class From_Ruby<NonlinearityType>
-  {
+  class From_Ruby<NonlinearityType> {
   public:
-    NonlinearityType convert(VALUE x)
-    {
+    From_Ruby() = default;
+
+    explicit From_Ruby(Arg* arg) : arg_(arg) { }
+
+    double is_convertible(VALUE value) { return Convertible::Exact; }
+
+    NonlinearityType convert(VALUE x) {
       auto s = String(x).str();
       if (s == "linear") {
         return torch::kLinear;
@@ -145,28 +152,34 @@ namespace Rice::detail
         throw std::runtime_error("Unsupported nonlinearity type: " + s);
       }
     }
+
+  private:
+    Arg* arg_ = nullptr;
   };
 
   template<>
-  struct Type<Scalar>
-  {
-    static bool verify()
-    {
-      return true;
-    }
+  struct Type<Scalar> {
+    static bool verify() { return true; }
   };
 
   template<>
-  class From_Ruby<Scalar>
-  {
+  class From_Ruby<Scalar> {
   public:
-    Scalar convert(VALUE x)
-    {
+    From_Ruby() = default;
+
+    explicit From_Ruby(Arg* arg) : arg_(arg) { }
+
+    double is_convertible(VALUE value) { return Convertible::Exact; }
+
+    Scalar convert(VALUE x) {
       if (FIXNUM_P(x)) {
         return torch::Scalar(From_Ruby<int64_t>().convert(x));
       } else {
         return torch::Scalar(From_Ruby<double>().convert(x));
       }
     }
+
+  private:
+    Arg* arg_ = nullptr;
   };
-}
+} // namespace Rice::detail