+++ lib/torch/nn/convnd.rb
@@ -1,7 +1,7 @@
 module Torch
   module NN
     class ConvNd < Module
-      attr_reader :in_channels, :out_channels, :kernel_size, :stride, :padding, :dilation, :transposed, :output_paddding, :groups, :padding_mode
+      attr_reader :in_channels, :out_channels, :kernel_size, :stride, :padding, :dilation, :transposed, :output_padding, :groups, :padding_mode
 
       def initialize(in_channels, out_channels, kernel_size, stride, padding, dilation, transposed, output_padding, groups, bias, padding_mode)
         super()